refactor: collapse tool class hierarchy from 4+ levels to 2 base classes (#32)
