+++ tools/simple/__init__.py
@@ -1,18 +1,13 @@
 """
 Simple tools for PAL MCP.
 
-Simple tools follow a basic request → AI model → response pattern.
-They inherit from SimpleTool which provides streamlined functionality
-for tools that don't need multi-step workflows.
-
-Available simple tools:
-- chat: General chat and collaborative thinking
-- consensus: Multi-perspective analysis
-- listmodels: Model listing and information
-- testgen: Test generation
-- tracer: Execution tracing
+Simple tools now inherit directly from BaseTool.
+This module provides backward compatibility.
 """
 
-from .base import SimpleTool
+from tools.shared.base_tool import BaseTool
+
+# Backward compatibility alias
+SimpleTool = BaseTool
 
 __all__ = ["SimpleTool"]
