+++ tools/thinkdeep.py
@@ -23,7 +23,7 @@ from shared_types import ToolModelCategory
 from systemprompts import THINKDEEP_PROMPT
 from tools.shared.base_models import WorkflowRequest
 
-from .workflow.base import WorkflowTool
+from .workflow.stateful_tool import StatefulTool
 
 logger = logging.getLogger(__name__)
 
@@ -92,7 +92,7 @@ class ThinkDeepWorkflowRequest(WorkflowRequest):
     )
 
 
-class ThinkDeepTool(WorkflowTool):
+class ThinkDeepTool(StatefulTool):
     """
     ThinkDeep Workflow Tool - Systematic Deep Thinking Analysis
